More contrast in speed icons of patch editor
The 16x16 icons of terrain speeds in the patch editor were difficult
for me to read.  Redraw the icons with more contrast, using green for
the fastest (0) and red for the slowest (7) terrains.  Do this in SVG
so there will be no question of where the source code of the font is.

speeds.svg has all the speeds in the same image.  After exporting to
bitmaps, it can be split and palettized with commands like these:

for speed in 0 1 2 3 4 5 6 7
do
  convert -crop 16x16+$((speed*16))+0 -page 16x16+0+0 speeds-small.png png8:speed${speed}-small.png
  convert -crop 48x48+$((speed*48))+0 -page 48x48+0+0 speeds.png png8:speed${speed}.png
done