HandleCheats: always restore the Lua stack
When calling the HandleCheats Lua function, restore the depth of the
Lua stack even if HandleCheats is undefined or not a function.  If
LuaCall reports a Lua error from within HandleCheats, don't output
anything more about it.  If HandleCheats returns too few or too many
values, end the debug output with a newline.