Pass options for commands of make.py via seperate "option=value" pairs.
The new command options format makes it more convenient to define optional
arguments without needing to escape code for the shell.
    
Example:
 ./make.py release cflags="-Wall" debug cflags="-Wall -Werror"
Will build a release version with -Wall passed to the compiler when building
objects. It will also build a debug version with -Wall and -Werror passed
to the compiler.
    
This new calling convention of make.py resembles more the traditional make
option passing mechanism.