Work around GL_GLEXT_VERSION redefinition in SDL
Old SDL1 contains a copy of glext.h which conflicts with the one
delivered by the new Mesa. That's actually an issue in the
distribution. See for example the fix in Redhat
https://bugzilla.redhat.com/show_bug.cgi?id=1662778 or the
work-around in Wargus https://github.com/Wargus/stratagus/issues/241.

As many distributions have the same build issue, add a workaround
to avoid including the SDL1 glext definition.

The build error looks like this:

 In file included from engine/include/video.h:34,
                 from engine/include/player.h:245,
                 from engine/network/network.cpp:227:
 /nix/store/28wgbaaagxnb0w78flq2jj6j3xnrpb1j-SDL-1.2.15-dev/include/SDL/SDL_opengl.h:116: error: "GL_GLEXT_VERSION" redefined [-Werror]
  116 | #define GL_GLEXT_VERSION 29
      |
 In file included from /nix/store/6l2z6ixw6hp1m2knfb2kvpncvsnlkmpw-libGL-1.4.0-dev/include/GL/gl.h:2050,
                 from /nix/store/28wgbaaagxnb0w78flq2jj6j3xnrpb1j-SDL-1.2.15-dev/include/SDL/SDL_opengl.h:46,
                 from engine/include/video.h:34,
                 from engine/include/player.h:245,
                 from engine/network/network.cpp:227:
 /nix/store/6l2z6ixw6hp1m2knfb2kvpncvsnlkmpw-libGL-1.4.0-dev/include/GL/glext.h:54: note: this is the location of the previous definition
   54 | #define GL_GLEXT_VERSION 20190805
      |

This also addresses the issue reported in bug #58753: "Compile gives
error with SDL libraries".