Fix fabricate.p/make.py "--keep" option
The "--keep" option had the "store_false" action which causes
the default to be True. And by default, fabricate was storing
the strace files locally with distinct names instead of
temporary files. Due to another bug, each strace file was deleted
after each command.

To fix the problem, the '--keep' action should be 'store_true' and
strace files should be removed when '--keep' option was not set
(aka false).