Work around debug build failure dereferencing type-punned pointer
In the network stack, a lot of pointers are cast which break strict
anti-aliasing rules. That broke debug build as warnings are treated
as errors.

This fix is a very crude work around. The network code is brittle
a serious rewrite of the network code would be needed.

Error output:

 engine/network/network.cpp: In member function ‘int CNetworkPacket::Deserialize(unsigned char*, unsigned int)’:
 engine/network/network.cpp:425:26: error: dereferencing type-punned pointer will break strict-aliasing rules [-Werror=strict-aliasing]
  425 |                         ((CNetworkExtendedCommand *)&this->Command[i])->Deserialize(p);