Remove broken and unmaintained SCons support
SCons was not used by members of the project since a decade. It was
not tested anymore and does not run anymore on new systems without
updates to the the SCons scripts.

The build system with make.py works better and exists since many
years, giving users plenty of time to try it out and adapt.