Make detection of libtheora more reliable
Make.py detected if libtheora was available by trying to link a
test executable with libtheora. This worked also for so-called
"static" builds where the linker could find libtheora.a. But
on Debian 11, the final link step failed because the Debian version
of libtheora.a relies on cairo which was not available as a static
lib.

Example error output:

  /usr/bin/ld: /usr/lib/gcc/x86_64-linux-gnu/10/../../../x86_64-linux-gnu/libtheora.a(decode.o): in function `th_decode_ycbcr_out':
(.text+0x7230): undefined reference to `cairo_image_surface_create'
  /usr/bin/ld: (.text+0x723d): undefined reference to `cairo_image_surface_get_data'
  /usr/bin/ld: (.text+0x7255): undefined reference to `cairo_image_surface_get_stride'
  /usr/bin/ld: (.text+0x74a6): undefined reference to `cairo_create'

To avoid that issue, the detection of libtheora also checks that it
can build with a call to a theora function.