In rectangle selection of a mix of mobile units, only select attackers
When a selection rectangle includes buildings and mobile units, only
mobile units are selected. In the same way, when there's a mix of
attackers and non-attackers, only select attacker units. This makes
it easier in the middle of battle to send troops to attack without
all the workers or engineers.