sdl: Fix text input (messages, edit boxes) with SDL2
With SDL2, SDL_KEYDOWN (and SDL_KEYUP) events give the key that was
pressed ignoring the modifiers. It was not possible to enter upper
case letters or alternate characters with the shift modifier.

In SDL1.2 'unicode' field in the event gave a character matching
the key modifiers (like shift) and the keyboard layout. That does
not exist in SDL2. Instead, one has to use SDL_TEXTINPUT.

This is an initial implementation to reach similar functionnality
as with SDL1.2. Ideally text input should be enabled and disabled
only when needed with SDL_StartTextInput() and
SDL_StopTextInput(). But the structure of the code and the
integration with Guichan made it too much for this commit.