sdl: fix changing screen size in the Video options menu
Before, when changing to a bigger size, the screen was pixelated
and only a portion of the screen was updated. When switching to a
smaller size, the display was bad again but only a portion of the
real screen was visible and accessible, some buttons went outside
of the screen.