sdl: fix duplicate keyboard events
With the migration to SDL2, text input and keyboard down events
can arrive for the same key. HandleKeyDown() has very special
(and hacky) logic to avoid doing the action twice when in
KeyInputState (for example writing a message). But it can also
cause issue outside text editing. That can break some keyboard
shortcuts like ALT+F for fullscreen. As that shortcut was seen
twice, the game toggled fullscreen twice coming back to the same
situation as if nothing happened.

To solve this, filter calls originating from TEXTINPUT events
when not in KeyStateInput.