Support default data path relative to the executable location.
To run, the Bos Wars engine needs to locate and load all the data
(scripts, graphics, sounds, units, menus, ...). The default path is
configured by the compile time variable STRATAGUS_LIB_PATH to look
in the current working directory.

That's brittle as the engine won't find the data when started from
another directory.

An alternative is to hardcode the absolute path. That works for
distributions, but is not practical for independent packages or
development.

Some packagers use wrapper scripts to work around this problem.
The script sets the data path via the "-d" command line option.
But that's not always practical.

To support such relative paths, the "~~" sequence at the start
of STRATAGUS_LIB_PATH will be replaced by the location of the
executable.