Merge user documentations from doc/ into README
The doc/ directory mixes a lot of different content:

1. Description & Basic User documentation
2. Build instructions
3. Licenses
4. Contribute & Development

But a lot of information is duplicated a bit everywhere. To keep
things clean, easier to access and manageable, the doc is being
re-organized. The README will contain 1 and 2. 3 will go into a
dedicated directory. 4 will be on the website and wiki, because
to contribute to the project, one needs to go online anyway.

Those is documentation on how to get and run the game. Building
and development is not moved here.