sdl: Grab (mouse input to) window instead of relative mouse mode
The initial migration to SDL2, used relative mouse mode for the
grab mode. But that's not ideal for tablets and touch inputs. SDL2
has new functions for this which should be used instead.

SDL_SetWindowGrab() and SDL_GetWindowGrab() instead of
SDL_SetWindowMouseGrab() and SDL_GetWindowMouseGrab() because those
are available only since SDL 2.0.16 and distros like Debian
oldstable are at SDL 2.0.14.