sdl: fix arrow keys broken in text input fields
When editing in a text input field (for example the nickname in
the Multiplayer menu or the filename in save game menu), keyboard
left <- and right -> arrow keys don't work to move the cursor.
One can only move the cursor with the mouse, which is inconvenient.
The backspace has the same issue.

This worked in version 2.8. It's a regression caused by the
migration to SDL2.