sdl: fix duplicate keyboard events in editors
With the migration to SDL2, text input and keyboard down events
can arrive for the same key. HandleKeyDown() has been fixed to
handle some keyboard shortcuts like ALT+F for fullscreen. As that
shortcut was seen twice, the game toggled fullscreen twice coming
back to the same situation as if nothing happened. The same issue
exists in the map editor and in the patch editor.

To solve this issue, filter calls originating from TEXTINPUT
events.