Reveal the map by default.
Hiding unexplored terrain forces the player to explore the map the
first time, which is boring. Players who alreaddy played the map
know the terrain already. Map previews also spoil the terrain.

Construction on unexeplored terrain is not allowed. Revealing the
map makes that issue go away.

In the future, we may support only revealed maps and drop
unexplored terrain.