engine: Accelerate display with SDL2 renderer
SDL2 has a renderer API to accelerate operations like drawing
lines, blitting graphics, drawing rectangles and so on.
That renderer can use 3D hardware to accelerate those, using
OpenGL or DirectX under the hood.

Converted:

 * rectangles (eventually with transparency)
 * lines (eventually with transparency)
 * image (graphics) blitting
 * filled rectangles (eventually with transparency)

The Minimap is generated in the classical way with a (software)
surface before rendering it via a texture.