engine: Accelerate drawing of fog tiles borders and explored
Explored terrain under fog was displayed with normal map on
top of which a black graphic was blended to create the grey
effect. For higher efficiency, this can be replaced by simply
drawing a black rectangle with intermediate transparency.

A special fog graphic was created for the border between foggy
explored terrain and visible terrain. The tiles of it were drawn
on top of the rendered map to create the fog effect. Instead of
creating an extra graphic, the normal fog border graphic is reused
directly and drawn with intermediate transparency.