menus: Run, load and save custom maps in user directory.
The editor now writes maps in the 'maps/' subdirectory of the user
directory (typically ~/.boswars/maps). Before, the editor saved maps
in the "maps/" subdirectory of the main data directory. But that
directory is sometimes read-only, for example in Linux distros.

Disrtibution with tools like Flatpaks and AppImage will also
benefit from this.

To make the newly created and saved maps from the editor playable,
it's now possible to run maps from the user directory.